api/requirements.txt
fastapi==0.96.0
uvicorn[standard]==0.22.0
gunicorn==20.1.0
pydantic==1.10.8
python-multipart==0.0.6
Cython==0.29.35
pandas==2.0.2
tiktoken==0.4.0
langchain==0.0.191
openai==0.27.8
faiss-cpu==1.7.4
nltk==3.8.1
torch==2.0.1
transformers==4.29.2
peft==0.3.0
optimum[neural-compressor]==1.8.7
neural-compressor==2.1.1
webvtt-py==0.4.6
intel_extension_for_pytorch==2.0.100
scikit-learn==1.2.2
scikit-learn-intelex==2023.1.1
